// src/interfaces/Report.ts
export interface IReportParam {
  users?: {} | any;
  query?: {} | any;
  params?: {} | any;
  body?: {} | any;
  type?: string;
  year?: string;
  month?: string;
  id?: String;
  full_name?: String;
  email?: String;
  phone?: String;
  instance?: String;
  visit_purpose?: String;
  page?: Number;
  offset?: Number;
}